+++ src/hal_common.c
@@ -169,11 +169,11 @@ extern int hal_fs_enable_chn_undistort(void *ctx, int chn);
 extern int hal_fs_disable_chn_undistort(void *ctx, int chn);
 
 /* ISP (hal_isp_gen1.c / hal_isp_gen2.c / hal_isp_gen3.c) */
-extern int hal_isp_set_brightness(void *ctx, uint8_t val);
-extern int hal_isp_set_contrast(void *ctx, uint8_t val);
-extern int hal_isp_set_saturation(void *ctx, uint8_t val);
-extern int hal_isp_set_sharpness(void *ctx, uint8_t val);
-extern int hal_isp_set_hue(void *ctx, uint8_t val);
+extern int hal_isp_set_brightness(void *ctx, int val);
+extern int hal_isp_set_contrast(void *ctx, int val);
+extern int hal_isp_set_saturation(void *ctx, int val);
+extern int hal_isp_set_sharpness(void *ctx, int val);
+extern int hal_isp_set_hue(void *ctx, int val);
 extern int hal_isp_set_hflip(void *ctx, int enable);
 extern int hal_isp_set_vflip(void *ctx, int enable);
 extern int hal_isp_set_running_mode(void *ctx, rss_isp_mode_t mode);
@@ -181,15 +181,15 @@ extern int hal_isp_set_sensor_fps(void *ctx, uint32_t fps_num, uint32_t fps_den)
 extern int hal_isp_set_antiflicker(void *ctx, rss_antiflicker_t mode);
 extern int hal_isp_set_wb(void *ctx, const rss_wb_config_t *wb_cfg);
 extern int hal_isp_get_exposure(void *ctx, rss_exposure_t *exposure);
-extern int hal_isp_set_sinter_strength(void *ctx, uint8_t val);
-extern int hal_isp_set_temper_strength(void *ctx, uint8_t val);
+extern int hal_isp_set_sinter_strength(void *ctx, int val);
+extern int hal_isp_set_temper_strength(void *ctx, int val);
 extern int hal_isp_set_defog(void *ctx, int enable);
-extern int hal_isp_set_dpc_strength(void *ctx, uint8_t val);
-extern int hal_isp_set_drc_strength(void *ctx, uint8_t val);
+extern int hal_isp_set_dpc_strength(void *ctx, int val);
+extern int hal_isp_set_drc_strength(void *ctx, int val);
 extern int hal_isp_set_ae_comp(void *ctx, int val);
-extern int hal_isp_set_max_again(void *ctx, uint32_t gain);
-extern int hal_isp_set_max_dgain(void *ctx, uint32_t gain);
-extern int hal_isp_set_highlight_depress(void *ctx, uint8_t val);
+extern int hal_isp_set_max_again(void *ctx, int gain);
+extern int hal_isp_set_max_dgain(void *ctx, int gain);
+extern int hal_isp_set_highlight_depress(void *ctx, int val);
 
 /* ISP getters and advanced tuning (hal_isp.c) */
 extern int hal_isp_get_brightness(void *ctx, uint8_t *val);
@@ -269,7 +269,7 @@ extern int hal_isp_get_awb_ct_attr(void *ctx, void *ct_attr);
 extern int hal_isp_set_awb_ct_attr(void *ctx, const void *ct_attr);
 extern int hal_isp_get_awb_ct_trend(void *ctx, void *trend);
 extern int hal_isp_set_awb_ct_trend(void *ctx, const void *trend);
-extern int hal_isp_set_backlight_comp(void *ctx, uint32_t strength);
+extern int hal_isp_set_backlight_comp(void *ctx, int strength);
 extern int hal_isp_get_defog_strength_adv(void *ctx, void *defog_attr);
 extern int hal_isp_set_defog_strength_adv(void *ctx, const void *defog_attr);
 extern int hal_isp_get_front_crop(void *ctx, void *crop_attr);
@@ -297,21 +297,21 @@ extern int hal_isp_get_wdr_output_mode(void *ctx, int *mode);
 extern int hal_isp_set_scaler_lv(void *ctx, int chn, int level);
 
 /* Multi-sensor ISP tuning (hal_isp.c) */
-extern int hal_isp_set_brightness_n(void *ctx, int sensor_idx, uint8_t val);
-extern int hal_isp_set_contrast_n(void *ctx, int sensor_idx, uint8_t val);
-extern int hal_isp_set_saturation_n(void *ctx, int sensor_idx, uint8_t val);
-extern int hal_isp_set_sharpness_n(void *ctx, int sensor_idx, uint8_t val);
-extern int hal_isp_set_hue_n(void *ctx, int sensor_idx, uint8_t val);
+extern int hal_isp_set_brightness_n(void *ctx, int sensor_idx, int val);
+extern int hal_isp_set_contrast_n(void *ctx, int sensor_idx, int val);
+extern int hal_isp_set_saturation_n(void *ctx, int sensor_idx, int val);
+extern int hal_isp_set_sharpness_n(void *ctx, int sensor_idx, int val);
+extern int hal_isp_set_hue_n(void *ctx, int sensor_idx, int val);
 extern int hal_isp_set_hflip_n(void *ctx, int sensor_idx, int enable);
 extern int hal_isp_set_vflip_n(void *ctx, int sensor_idx, int enable);
 extern int hal_isp_set_running_mode_n(void *ctx, int sensor_idx, rss_isp_mode_t mode);
 extern int hal_isp_set_sensor_fps_n(void *ctx, int sensor_idx, uint32_t fps_num, uint32_t fps_den);
 extern int hal_isp_set_antiflicker_n(void *ctx, int sensor_idx, rss_antiflicker_t mode);
-extern int hal_isp_set_sinter_strength_n(void *ctx, int sensor_idx, uint8_t val);
-extern int hal_isp_set_temper_strength_n(void *ctx, int sensor_idx, uint8_t val);
+extern int hal_isp_set_sinter_strength_n(void *ctx, int sensor_idx, int val);
+extern int hal_isp_set_temper_strength_n(void *ctx, int sensor_idx, int val);
 extern int hal_isp_set_ae_comp_n(void *ctx, int sensor_idx, int val);
-extern int hal_isp_set_max_again_n(void *ctx, int sensor_idx, uint32_t gain);
-extern int hal_isp_set_max_dgain_n(void *ctx, int sensor_idx, uint32_t gain);
+extern int hal_isp_set_max_again_n(void *ctx, int sensor_idx, int gain);
+extern int hal_isp_set_max_dgain_n(void *ctx, int sensor_idx, int gain);
 extern int hal_isp_get_exposure_n(void *ctx, int sensor_idx, rss_exposure_t *exposure);
 extern int hal_isp_set_custom_mode_n(void *ctx, int sensor_idx, int mode);
 extern int hal_isp_set_ae_freeze_n(void *ctx, int sensor_idx, int enable);
@@ -449,8 +449,18 @@ extern int hal_ivs_destroy_base_move_interface(void *ctx, void *handle);
 extern void *hal_ivs_create_persondet_interface(void *ctx, void *param);
 extern int hal_ivs_destroy_persondet_interface(void *ctx, void *handle);
 /* JZDL uses standalone API (hal_jzdl_create/detect/destroy), not IVS interface */
-static void *hal_ivs_create_jzdl_stub(void *ctx, void *param) { (void)ctx; (void)param; return NULL; }
-static int hal_ivs_destroy_jzdl_stub(void *ctx, void *handle) { (void)ctx; (void)handle; return 0; }
+static void *hal_ivs_create_jzdl_stub(void *ctx, void *param)
+{
+    (void)ctx;
+    (void)param;
+    return NULL;
+}
+static int hal_ivs_destroy_jzdl_stub(void *ctx, void *handle)
+{
+    (void)ctx;
+    (void)handle;
+    return 0;
+}
 
 #endif /* HAL_MODULE_VIDEO */
